Move expert routes to expertRoutes.js
