+++ backend/Routes/auth.js
@@ -2,3 +2,4 @@
 router.get('/test', (req, res) => {
   res.json({ message: 'Auth endpoint is working' });
 });
+
